--- lib/ui/barcosDTO.dart
@@ -1,11 +1,0 @@
-
-class barcosDTO {
-  String IDBarco = '';
-  String foto1 = '';
-  String foto2 = '';
-  String foto3 = '';
-  String foto4 = '';
-  String foto5 = '';
-  String nomeBarco = '';
-  int tamanho = 0;
-}
