+++ lib/ui/barcos_dto.dart
@@ -1,0 +1,12 @@
+// ignore_for_file: avoid_print
+
+class BarcosDTO {
+  String iDBarco = '';
+  String foto1 = '';
+  String foto2 = '';
+  String foto3 = '';
+  String foto4 = '';
+  String foto5 = '';
+  String nomeBarco = '';
+  int tamanho = 0;
+}
