--- lib/ui/partidaAleatoriaDTO.dart
@@ -1,16 +1,0 @@
-import 'package:proj0511/ui/barcosDTO.dart';
-
-
-class partidaAleatoriaDTO {
-  String? elo;
-  String? idAdversario;
-  String? nacionalidade;
-  String? nomeAdversario;
-  List<barcosDTO> barcos = [];
-  String idPartida = '';
-  String? foto;
-  String? idCenario;
-  String? nomeCenario;
-  String? descricaoCenario;
-  String? message;
-}
