+++ lib/ui/partida_aleatoria_dto.dart
@@ -1,0 +1,15 @@
+import 'package:proj0511/ui/barcos_dto.dart';
+
+class partidaAleatoriaDTO {
+  String? elo;
+  String? idAdversario;
+  String? nacionalidade;
+  String? nomeAdversario;
+  List<BarcosDTO> barcos = [];
+  String idPartida = '';
+  String? foto;
+  String? idCenario;
+  String? nomeCenario;
+  String? descricaoCenario;
+  String? message;
+}
