+++ lib/ui/validador.dart
@@ -1,19 +1,13 @@
 import 'package:flutter/material.dart';
 
-bool emailValid(String email){
+bool emailValid(String email) {
   final RegExp regex = RegExp(
       r"^(([^<>()[\]\\.,;:\s@\']+(\.[^<>()[\]\\.,;:\s@\']+)*)|(\'.+\'))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$");
   return regex.hasMatch(email);
 }
 
-bool passValid(String pass){
+bool passValid(String pass) {
   final RegExp regex = RegExp(
       r"^(?=.*[@!#$%^&*()/\\])(?=.*[0-9])(?=.*[a-zA-Z])[@!#$%^&*()/\\a-zA-Z0-9]{6,50}$");
   return regex.hasMatch(pass);
 }
-
-
-
-
-
-
